requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
PyJWT==2.1.0
six==1.16.0
Werkzeug==2.0.1
